requirements.txt
# Streamlit UI
streamlit>=1.29.0

# Vector database
chromadb>=0.4.0

# LLM frameworks
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# LLM providers
ollama>=0.3.0
openai>=1.0.0
anthropic>=0.8.0

# OCR and image processing
pytesseract>=0.3.10
pillow>=10.0.0

# Screenshot capture
pyautogui>=0.9.54

# PDF processing
pypdf2>=3.0.0

# Embeddings
sentence-transformers>=2.2.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
